backend/requirements.txt
# backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.1
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-bdd==6.1.1
pytest-asyncio==0.21.1
alembic==1.12.1
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
